[#34] 네비게이션 바 구조 수정
- 네비게이션 컴포넌트 구조 분리 (Desktop/Tablet)
- 태블릿 환경에서 아이콘만 표시되도록 수정
- 스타일 코드 모듈화 및 컴포넌트 내부로 이동

해당 내용을 통해
1. 컴포넌트 구조 개선
2. 코드 모듈화
를 하였습니다.
